# Optimize TRON sync