before this PR, the TRON sync logic have no "reconciliation" mechanism and it completely resync everything from scratch. This has shown to be not optimized especially on one part: the gettransactionbyid calls, we seem to need to fetch each individual tx at the moment which is very bad for performance. The solution provided here store in the account a cache that avoid fetching again for the second synchronisation. **it still makes the first sync really slow and it's something we absolutely need to address by backend**

To test and measure that, i've used one of the biggest public account, this one only have 152 transactions but it still take 52 seconds to sync the first time. the optim will make the second sync go to 9s. 

```sh
VERBOSE_FILE=trace BRIDGE_FORCE_IMPLEMENTATION=js ledger-live sync -c tron --xpub TKTWiCfNyitRiA31eiMLTTFYah6m2aLkup --format json > account.json

# first sync takes 52s

ledger-live sync -c tron --file account.json --format summary

# second sync now takes 9s   (used to be 52s)
```

It stores in the account raw a compressed cache object, that looks like this:

```js
"3a8afc3f9052dcd6c660453a20af6d98a0c81fde7251fc008d2c3403bbde01ba":[null,13562675,null],"bfd1053548fbcc903ead63746006749911bc7be56ed6a45bb3c32d22a54a768d":[null,13540030,null],"bb96ac10558c1cbc0a031387aafc541670ea4f6f3092f3ade341752728aa3ce3":[null,13548724,null],"4b27cc4afe55d855819daea4a90ebedc1d8cf55bb6603a0951f53f1fd8b1a6eb":[null,13534657,null],"fab086178c65fa2aa868cc487c025a8503492adfe6ef0d4560785de254203926":[null,13539961,null],"30665688d23bee0b8429ae33e0caf0fd21000f5d1fb05dcdeb0c87cc7cf12240":[null,13515774,null],"938712a111f1f8ebd37f224e9fb33bcc44ba076f67da150c270b659be6328902":...
```

for this examlpe shared about, the memory of the JSON data is an addition of **66K**. negligible.

## fetch by 100

the second optim is to move away from default of fetch by 20, but to fetch by 100 transactions. we might even go higher in future.